ai-agent/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
redis==4.6.0
google-generativeai==0.3.0
pydantic==2.4.2
